Correctly handle not having gis installed
When we handle not being able to import the gis Point type when gis is
not installed we don't set up any value for the Point variable. This
means later when we try and test type it is not defined and we get a
NameError.

Set Point to None and test if it is set later.